src/measurements-collector/requirements.txt
ptyprocess==0.7.0
reactivex==4.0.4
ruuvitag-sensor==2.0.0
typing-extensions==4.3.0
##
python-jose==3.3.0
python-keycloak==2.9.0
requests==2.28.2
urllib3==1.26.14